# Use _div64 intrinsic if available