Following the theme of #1769, use the [_div64](https://learn.microsoft.com/en-us/cpp/intrinsics/div64?view=msvc-170) intrinsic to  have FixedDiv perform a 32-bit operand divide like the original Doom. This is available since VS 2019 and is worth about a 1% faster tic (*) and a 1.3x+ speedup when using voxels.

(*) Intel CPUs newer than Coffee Lake have a hardware divider and won't see any benefit from this PR

|| fps before [realtics] | fps after [realtics]|
|--------|--------|--------|
| [av20](https://dsdarchive.com/files/demos/av/28588/av20-1357.zip) | 2981 [345] | 2999 [343] |
| [su03](https://dsdarchive.com/files/demos/sunder/56326/su03-935.zip) | 1259 [565] | 1270 [560] |
| [coma](https://dsdarchive.com/files/demos/comatose/59073/comatose01p040.zip) | 141 [405] | 142 [401] |

with voxels:

|| fps before [realtics] | fps after [realtics]|
|--------|--------|--------|
| av20 | 1495 [688] | 1941 [530] |
| su03 | 289 [2465] | 400 [1799] |

Testing was done with `-nosound -noblit` and voxel tests with `-fastdemo` to speed things up a bit.