@rfomin 2024-07-13
I can confirm, nice speed improvements with voxels on my old laptop. I also made a GCC version:
```diff
diff --git a/src/m_fixed.h b/src/m_fixed.h
index 91d43eab..33b00667 100644
--- a/src/m_fixed.h
+++ b/src/m_fixed.h
@@ -24,12 +24,40 @@
 #include <stdlib.h> // abs()
 #include <stdint.h> // int64_t
 
-// VS 2019 and later provide a [64]/[32] intrinsic
-#if defined(_MSC_VER) && _MSC_VER >= 1920 && (defined(_M_X64) || defined(_M_IX86))
-  #include <immintrin.h>
-  #define DIV64(a,b) (_div64((a),(b),NULL))
+#include "doomtype.h"
+
+#if defined(_M_X64) || defined(_M_IX86)
+
+  #if defined(_MSC_VER) && _MSC_VER >= 1920
+    #include <immintrin.h>
+  #endif
+
+  inline static int32_t div64_32(int64_t a, int32_t b)
+  {
+  // VS 2019 and later provide a [64]/[32] intrinsic
+  #if defined(_MSC_VER) && _MSC_VER >= 1920
+      return _div64(a, b, NULL);
+  #elif defined(__GNUC__)
+      if (__builtin_constant_p(b))
+      {
+          return a / b;
+      }
+      int32_t lo = a;
+      int32_t hi = a >> 32;
+      __asm__("idivl %2" : "+a"(lo), "+d"(hi) : "rm"(b));
+      return lo;
+  #else
+      return a / b;
+  #endif
+  }
+
 #else
-  #define DIV64(a,b) ((fixed_t)((a)/(b)))
+
+  inline static int32_t div64_32(int64_t a, int32_t b)
+  {
+      return a / b;
+  }
+
 #endif
 
 //
@@ -65,7 +93,7 @@ inline static fixed_t FixedDiv(fixed_t a, fixed_t b)
 {
   // [FG] avoid 31-bit shift (from Chocolate Doom)
   return (abs(a)>>14) >= abs(b) ? ((a^b) < 0 ? INT_MIN : INT_MAX) :
-    DIV64((int64_t) a << FRACBITS, b);
+    div64_32((int64_t) a << FRACBITS, b);
 }
 
 #endif
```
Tested with MSYS.