@rfomin 2024-07-13
> Should we move the feature check for the presence of `_div64()` to the CMake build system?

Yes, I think so. I can do it later.