@gendlin 2024-07-13
I found one problem case with INT_MIN / -1, which silently returns 0 in the existing FixedDiv but is a hardware overflow exception in the new code. I can add a check for this, unless it's just something that should never happen or should actually cause a crash?