@rfomin 2024-07-13
> I found one problem case with INT_MIN / -1, which silently returns 0 in the existing FixedDiv but is a hardware overflow exception in the new code.

Interestingly, the GCC version works fine (silently returns 0). I think we need to add check or maybe write an asm-version for MSVC.