@gendlin 2024-07-13
> GCC version works fine

It seems like you should be getting an exception - are you seeing the perf improvements with the gcc version? I added the check in my version and pushed some test code [here](https://github.com/gendlin/woof/commit/c1556fa97d935e5967f1f775bb5087ea29e97a0f)