@rfomin 2024-07-13
> It seems like you should be getting an exception - are you seeing the perf improvements with the gcc version? I added the check in my version and pushed some test code [here](https://github.com/gendlin/woof/commit/c1556fa97d935e5967f1f775bb5087ea29e97a0f)

Yes, I got the same performance gain with voxels as the MSVC version. I ran your test successfully (although I didn't do full UINT_MAX iterations). Try it yourself - it's pretty easy to set up GCC/Clang using https://www.msys2.org/ or WSL.