@gendlin 2024-07-13
Ok, interesting. I guess the exceptions are just masked differently or the emitted assembly is not quite the same