@rfomin 2024-07-13
@gendlin Wait, I didn't get a crash before because of this check:
```c
if (__builtin_constant_p(b))
{
    return a / b;
}
```
`__builtin_constant_p` checks if expression is known to be constant at comile time. If I remove this check, it also fails at `(INT_MIN, -1)`.