@gendlin 2024-07-13
Ah yeah, good catch. Might as well just remove that check? I don't think there are many FixedDiv calls with constant args